Add directional indicator to p-value heatmap
Add a directional indicator to the p-value heatmap. When comparing two
groups, the Levene's test p-value determines whether the two variances
are equal. However, the p-value does not indicate which group has the
larger variance.

The indicator takes the form of a small circle plotted within each cell
of the p-value heatmap. If on the top left corner, the group on the row
has the larger variance. If on the bottom right corner, the group on the
column has the larger variance.

The direction of each indicator is calculated by the stats module, and
the plotting of the circles is done in the plots module.

The pairwise Levene's test calculator was rewritten to return a
dictionary instead of a square dataframe. This makes hypothesis testing
easier since it becomes easier to access the p values.