Analyze drift for full worm transcriptome
Analyze the distribution of CPM and drift for the full and smaller
subset of the worm transcriptome.

Also removed the plot of drift for mianserin added on different days for
the subset of the transcriptome because it was not too informative.

It seems that sample 5 (50 uM mianserin added on day 1, harvested on day
3) has a drift distribution that is very different from the rest of the
samples. The drift values are highly centered around 0 in both the full
and smaller transcriptome data.