Modify Gene Ontology hierarchy traversal function
Remove the old method for determining which genes belong to which Gene
Ontology terms. The old method would perform a O(N^2) traversal of the
DAG to generate a list of nodes which can reach a particular node. This
information was then used for a O(N^2) algorithm that unions together
the genes in the nodes which can reach the current node to find the gene
set for a GO term.

Instead, the new DFS is O(N) with memoization of genes for each GO term.
The DAG has been converted into a top down tree with three root nodes.
No intermediate representation is generated, and the return result is
just a set of genes associated with each GO term according to the
ontology. Dramatically improves runtime (1 min to 2 s).