Update GO term drift analysis with hierarchy in GSE69244
Update the original preliminary analysis of drift grouped by GO term for
GSE69244 with the GO hierarchy taken into consideration. The code now
looks at not only the GO terms used in annotations, but also at terms
higher up the ontology that encapsulate the annotation terms.

Results in an increased number of GO terms with a testable number of
genes. Finds 95 GO terms with significant reduction in drift when given
J147, but most have a very large number of genes.