Update mianserin analysis with GO hierarchy grouping
Update analysis of drift in GO groups for the mianserin worm dataset
with the GO hierarchy taken into account.

Since more GO groups are tested, we find an increased number of GO
groups with significant drift and reduction in drift when given
mianserin. However, none of these groups seem to overlap with those
found in the analysis of the mice data.

Overall at the moment it is still difficult to determine which GO groups
are interesting, since there are so many of them. In addition, we will
need a way to determine whether GO groups with hundreds or thousands of
genes are interesting at all, since they seem to always drift.