Create Gaussian variance detection simulation
Create preliminary analysis of the probability of detecting differences
in variance between two Gaussian distributions with Bartlett's test.

We want to know how much of a difference in variance is detectable with
certain sample sizes. This allows us to make informed decisions about
the observed differences in drift that are being measured with RNA-seq.

This is a proof-of-concept with the Gaussian distribution. The results
show that at large sample sizes, even small differences in variance
between two distributions can be detected.