Increase flexibility of multi density plot function
Change the multi density plot function to plot multiple times on the
same Matplotlib axes with the Seaborn kdeplot function. The previous
version used a pandas pivot call to reshape the dataframe prior to
plotting, which was very slow.

The old version also had to have the same number of observations for
each group, which would give problems with NaNs after pivoting.

This change significantly sped up the function. Plotting ~19000
observations for 12 groups went from ~7 seconds to ~800 ms.