Analyze influence of gene expression level on drift
Perform initial analysis of how gene expression affects drift for the
filtered transcriptome of PMID26623667.

The motivating hypothesis is that genes which are originally very lowly
expressed will drift differently from genes which were highly expressed
in the young reference.

Results: it seems that for the filtered transcriptome the lowly
expressed genes do skew towards having positive drift much more than the
rest of the transcriptome. This is likely a selection bias due to the
drift statistic, since genes which reach zero expression cannot be used
to calculate drift.

Another surprising result is that the drift variance of genes also
changes depending on the original gene expression level. The lowest and
highest deciles of genes have much larger drift variances compared to
the middle 80% of genes. This can be seen across all samples.