Create metadata parser for GSE77110
Create notebook to extract sample metadata for GSE77110.

This experiment has a relatively simple design. There is one young
reference (normal diet, day 2). For each of the three dieting conditions
(control, caloric restriction, intermittent fasting), RNA was harvested
on days 4, 6, 8, and 10. RNA was also harvested on days 12 and 14 for
the intermittent fasting condition.