Create preliminary analysis of GSE77110
Add exploratory analysis of GSE77110. The strange thing with this
dataset is that there only seems to be one biological replicate for each
time point and condition, which removes any statistical power to say
that the drift variance is different between two conditions.

Will still likely be useful for determining a set of genes which
accurately reflect the population variance of the log fold expression
change across multiple conditions.