Create worm behaviour phenotype file for GSE12290
The public GSE12290 data for PMID18778409 did not contain information
about the movement phenotype of each C. elegans worm.

I emailed Dr. Simon Melov requesting these data, and he replied with an
Excel spreadsheet with the mapping. I changed this to a tsv and have
included it here.

Email was received from Dr. Melov on 2016-05-13.