Create initial analysis of GSE12290
First preliminary look at the data in GSE12290. Seems like the provided
expression values are relative to pooled worms on day 4, since the
distribution of the values have a mean of zero and are symmetric about
the mean.

Overall there does seem to be some correlation between drift and time,
but in these data the correlation seems weak. I see no obvious
differences between the different phenotyped worms either.