Create initial drift regressions for PMID26623667
The data presented in the paper showed that the variance in the relative
log fold expression change increased with time, and that by Levene's
test is statistically significant between two time points. From this the
authors concluded that drift is increasing with time, and that drift for
older mianserin treated worms was the same as that of younger control
worms.

This second claim was bothering me, since I felt that Levene's test was
not the correct statistical test to use to prove this assertion. Rather
I felt that it was necessary to show that the mean drift value across
multiple biological replicates was different between the two groups.

This notebook demonstrates that the drift is indeed lower in mianserin
treated worms. By plotting drift for each biological sample, we see a
clear increasing trend that seems to be best approximated with a
log-linear model. I feel that the linear regression plots in this
notebook should have been included in the paper to support the assertion
that drift increases with time, and is reduced in mianserin treated
worms.

At the moment it seems that Levene's test will be used less in the
future, since it is not actually comparing drift, but rather the
distributions used to calculate drift for a single sample.