Remove probe mappings from GSE21784 preprocessor
Refactor the GSE21784 data preprocessor to no longer annotate probes
with the WormBase and Entrez gene identifiers. Due to difficulties with
identifier conversion (missing mappings, multiple mappings, conflicting
mappings), the logic for how to subset and filter probes is left to the
analysis portion of the pipeline.

The updated data preprocessor is only responsible for determining the
relevant metadata for each sample, reshaping the expression matrix to
long form, and for ensuring that the sample metadata is joined together
with the expression values.

Results:
All 9 biological samples contain expression values for 22625 probes.