Rewrite GEO parser to reshape expression matrix
Rewrite GEO series matrix parser to reshape the expression matrix to
long format. Also now checks that no values in the expression matrix are
empty, and that the GSM ids are correctly formatted.

Reshaping the matrix from wide form to long form was a common step for
the GSE data preparation scripts.