Fit linear model for drift in PMID26623667
Update initial analysis of drift in PMID26623667 with R code
demonstrating that the outcome of drift is significantly modeled by a
linear combination of the log of the day harvested and the drug
treatment. The prior analysis plotted the data to show that a
correlation existed, but now the actual model has been generated.

Results:
When drift is modeled as:
    drift ~ log(day_harvested) + treatment
the adjusted R-squared value is 0.9336. The interaction term between age
and drug is not significant, suggesting that the drug is equally
effective at reducing the drift at all time points examined.

The control condition is water and the treatment condition is mianserin.
The treatment label for "water" was changed to "control" to ensure that
the linear model calculates the effect of mianserin with respect to the
control and not vice versa.