Update initial GO drift analysis for PMID26623667
Replace the very initial analysis of identifying GO terms which drift
over time in PMID26623667 (worm_gene_ont.ipynb). The very first attempt
at identifying these GO terms was incorrect because it used Levene's
test to compare the variances of two relative log fold expression change
distributions. Levene's test can identify that the variances of the two
distributions are different, but it is not able to identify whether the
drift values are statistically different across multiple biological
replicates. Therefore we were using the wrong statistical test, making
the inferences incorrect.

The updated analysis is a refinement, but is still not correct since it
uses three t-tests to identify whether there is a reduction in drift
when mianserin is given. Instead of fitting two linear models
independently to the treatment and control groups, one model should be
fit which includes the treatment as another independent categorical
variable.