Separate GO drift obfuscation from drift analysis
Separate out the GO term drift obfuscation for Michael from the rest of
the GO term drift analysis. Keeping the obfuscation portion separate
will make it less likely to confuse the results.

The current obfuscation method was to simply label the good results with
the GO term metadata of less good results in order to preserve the fact
that similar GO terms may show up near each other due to shared genes.
Not too much effort was spent on intelligently obfuscating the results
because it is more important to move forwards with resequencing of
specific GO terms.