Analyze randomly generated gene exp.set. for drift
Examine the randomly generated gene expression dataset for drift.

Motivation:
Wanted to confirm that the drift pattern seen for the entire
transcriptome in PMID26623667 is not a property that would be expected
for all datasets.

Results:
The randomly generated dataset containing 20000 genes exhibits no trend
in drift with time. This is good, because it means that the drift trend
is not just a consequence of measuring the variance for 20000 different
gene expression values.