Create random gene sets of various sizes
Create a random sample of gene sets of various sizes (10 to 500 genes)
from the data for PMID26623667. Genes are randomly chosen from the 10879
genes with relative log-fold expression change values, and are
identified by their WormBase identifier.

All generated gene sets are stored in the genesets/ subfolder.

Two groups of random gene sets are created.
  1. uniform.json: This file contains 100 gene sets per gene set size,
    with gene set sizes ranging from 10 to 500 in increments of 1. Used
    for an overview analysis exploring the effect of gene set size on
    model parameters.
  2. size_*.json: These files contain 10000 gene sets per gene set size,
    with gene set sizes increasing in increments of 10. Intended for use
    as background distributions for GO term analysis.

The actual files are not included in the git history due to their large
size, but should be the same when generated each time.