Fit drift models for randomly generated gene sets
Calculate linear and log-linear drift models for the randomly generated
gene sets using PMID26623667.

Code was run in parallel on a 40-core Amazon Web Services node running
Ubuntu 14.04 LTS with Python 3.4.3 and R 3.3.0. Total runtime is
approximately 40 minutes.

Generated models are not stored in the repository due to space
constraints.