Analyze effect of gene set size on drift models
Create overview analysis of the influence of gene set size on drift
model parameters for the randomly generated gene sets of PMID26623667.

Conclusions:
Gene set size plays an important role in determining whether to consider
a gene set to be a significant model for drift. In the PMID26623667
dataset, a gene set is more likely to have significant model parameters
when the set contains many genes. Therefore, gene sets need to be
compared to a size-matched reference random distribution in order to
evaluate whether the specific gene set is important.