Re-analyze GO term drift analysis of PMID26623667
The original GO term drift analysis of PMID26623667 was done without any
real way of determining which results were actually significant, and
therefore should be re-examined by further sequencing. This was because
there was no information about the background distribution; i.e., how
well a randomly generated gene set would compare to the gene sets
created using the Gene Ontology.

The analysis of the drift of genes grouped by GO term has been repeated
here, but the results are now compared to randomly generated gene sets
of the same size. This allows the results for each GO term to be
considered in context.

Results:
Based on the overview of the randomly generated gene sets, we knew that
the adjusted R-squared of the models would increase as the number of
genes increased. This we believe is because the entire transcriptome
fits well to a linear model, and the larger gene sets are just better
approximations of the entire transcriptome.

When the GO terms are compared to the randomly generated gene sets, we
see that most of the larger GO terms do not score very well, since they
perform no better than what we would expect on average for random gene
sets of the same size. Most of the GO terms which perform better than
average are smaller in size. Interestingly, the GO terms which perform
best seem to be involved in the immune response.

Conclusions:
We now have a list of genes that would be interesting to re-sequence and
analyze further in greater detail.