Analyze proposed drift seq gene sets for quality
Examine the gene sets proposed by Michael for resequencing for how they
drift with time. The log-linear models of drift for each gene set are
compared against the background distribution of random gene sets in
order to contextualize the performance of each gene set.

Results:
Most of the gene sets have very poor drift model characteristics, and
would likely not be good indicators of drift. However, some of these
models were purposely chosen by Michael to act as controls, and
therefore having them not drift with time is a valuable property.

It is up to Michael to decide if he wants to sequence any of these
genes, since some are chosen to try and fit an aging narrative and not
based on whether the drift model performed exceptionally well.