Add proposed drift seq gene sets
Add the initial proposals for gene sets to re-sequence via drift-seq.
This list of gene sets was hand picked by Michael based on the earlier
GO term analyses that I performed.

We are trying to select a subset of genes to resequence with a greater
number of biological replicates and time points. We ideally would like
the drift profiles of these genes to match what was observed in
PMID26623667 to give us the highest chance of success when these genes
alone are resequenced in a greater number of samples.