Display GO term drift results in Cytoscape
Convert the results of the GO term drift re-analysis from a tabular
format to a visual network format. Previously it was difficult to
determine the relationship of GO terms and how many genes they shared in
common. Now the results are presented visually via Cytoscape 3.4, and
the relations between GO terms are more clearly visible.

To simplify the resulting visualization, only the top 15% of scored GO
terms and their children up to two edges away are displayed. This
reduces the amount of visual clutter, since only ~300 nodes are
displayed instead of ~1800.